Fix play/pause icons being offset from center.
They were made before I knew the icon paths needed to be aligned from
the (0,0) origin inside the svg file to be centered. The paths don't
have a view box, so mpv's osd has no way of telling if the icon
does not stop at the right edge of the play triangle tip at x=30
instead stopping at x=44 in the middle of the whitespace.