Disable playlist thumbnails
Since the mouse stays on the button while the next video loads, it
can cause mpv to generate a preview at the same time it's loading the
video. Since you wait a few seconds the for UI to unfreeze, it also
tries to show the thumbnail for the video after.

Playlist thumbnails need to be pre generated and appear instantly.
There should be a reasonable delay after pl_next is pressed before
showing the next thumbnail.