Parse keybindings for tooltips from input-bindings property (#5)
Mpv's stats.lua has a great example of parsing the bindings. We use
regex on the command to find all alternative shortcuts and list them
in the generated labels.

A number of obvious shortcuts like media keys are filtered out.

Prettified a few keys, Arrow keys are replaced with unicode arrows.